Delete presentation folder and refactor UI packages
