+++ app/src/androidTest/java/mx/yellowme/youst/challenges/navigation/GreetingFragmentTest.kt
@@ -1,4 +1,4 @@
-package mx.yellowme.youst.presentation.challenges.navigation
+package mx.yellowme.youst.challenges.navigation
 
 import android.os.Bundle
 import androidx.fragment.app.testing.launchFragmentInContainer
