+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/SharedAppTheme">
         <activity
-            android:name=".presentation.splash.SplashActivity"
+            android:name="mx.yellowme.youst.splash.SplashActivity"
             android:theme="@style/SharedAppTheme.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -21,7 +21,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".presentation.showcase.ShowcaseActivity"
+            android:name="mx.yellowme.youst.showcase.ShowcaseActivity"
             android:label="@string/app_name"
             android:theme="@style/SharedAppTheme.NoActionBar">
             <intent-filter>
@@ -40,29 +40,29 @@
             android:label="@string/pokemon_detail"
             android:theme="@style/SharedAppTheme.NoActionBar" />
         <activity
-            android:name=".presentation.challenges.CrazyListsChallengeActivity"
+            android:name="mx.yellowme.youst.challenges.CrazyListsChallengeActivity"
             android:label="@string/activity_challenge_title_crazy_lists"
             android:theme="@style/SharedAppTheme.NoActionBar" />
         <activity
-            android:name=".presentation.challenges.ListenToMeChallengeActivity"
+            android:name="mx.yellowme.youst.challenges.ListenToMeChallengeActivity"
             android:label="@string/activity_challenge_title_listen_to_me"
             android:theme="@style/SharedAppTheme.NoActionBar" />
         <activity
-            android:name=".presentation.challenges.navigation.NavigationActivity"
+            android:name="mx.yellowme.youst.challenges.navigation.NavigationActivity"
             android:label="Jetpack Navigation"
-            android:parentActivityName=".presentation.showcase.ShowcaseActivity"
+            android:parentActivityName="mx.yellowme.youst.showcase.ShowcaseActivity"
             android:theme="@style/SharedAppTheme.NoActionBar.TransparentStatus">
 
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".presentation.showcase.ShowcaseActivity" />
+                android:value="mx.yellowme.youst.showcase.ShowcaseActivity" />
         </activity>
 
         <meta-data
             android:name="preloaded_fonts"
             android:resource="@array/preloaded_fonts" />
 
-        <receiver android:name=".presentation.challenges.navigation.DeepLinkWidgetProvider">
+        <receiver android:name="mx.yellowme.youst.challenges.navigation.DeepLinkWidgetProvider">
             <intent-filter>
                 <action android:name="android.appwidget.action.APPWIDGET_UPDATE" />
             </intent-filter>
