+++ app/src/main/java/mx/yellowme/youst/challenges/BaseChallengeActivity.kt
@@ -1,4 +1,4 @@
-package mx.yellowme.youst.presentation.challenges
+package mx.yellowme.youst.challenges
 
 import android.os.Bundle
 import android.view.Menu
@@ -11,7 +11,7 @@ import mx.yellowme.youst.core.components.dialogs.SimpleInfoDialogBuilder
 import mx.yellowme.youst.core.domain.Challenge
 import mx.yellowme.youst.core.extensions.toast
 import mx.yellowme.youst.core.hooks.BaseActivity
-import mx.yellowme.youst.data.ChallengeDataHelper
+import mx.yellowme.youst.core.data.ChallengeDataHelper
 import mx.yellowme.youst.core.utils.findOrThrow
 
 /**
