+++ app/src/main/java/mx/yellowme/youst/pokemon/PokemonListLoader.kt
@@ -1,9 +1,11 @@
-package mx.yellowme.youst.data;
+package mx.yellowme.youst.pokemon;
 
-import mx.yellowme.youst.presentation.challenges.ChallengeWithListActivity
+import mx.yellowme.youst.core.data.APIServiceGenerator
+import mx.yellowme.youst.core.data.PokemonAPIService
+import mx.yellowme.youst.core.domain.Pokemon
 import mx.yellowme.youst.core.extensions.toast
 import mx.yellowme.youst.pokemon.index.PokemonListViewHolder
-import mx.yellowme.youst.core.domain.Pokemon
+import mx.yellowme.youst.challenges.ChallengeWithListActivity
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -27,9 +29,7 @@ object PokemonListLoader {
                 }
             }
 
-            override fun onFailure(
-                    call: Call<MutableList<Pokemon>>, t: Throwable
-            ) {
+            override fun onFailure(call: Call<MutableList<Pokemon>>, t: Throwable) {
                 view.setState(ChallengeWithListActivity.ListState.LOADED)
                 view.toast("Error")
             }
