+++ app/src/main/java/mx/yellowme/youst/showcase/cards/CardAdapter.kt
@@ -1,4 +1,4 @@
-package mx.yellowme.youst.presentation.showcase.cards
+package mx.yellowme.youst.showcase.cards
 
 import androidx.cardview.widget.CardView
