+++ app/src/main/java/mx/yellowme/youst/showcase/cards/CardPagerAdapter.kt
@@ -1,4 +1,4 @@
-package mx.yellowme.youst.presentation.showcase.cards;
+package mx.yellowme.youst.showcase.cards;
 
 import android.view.LayoutInflater
 import android.view.View
@@ -10,7 +10,7 @@ import androidx.viewpager.widget.PagerAdapter
 import com.google.android.material.button.MaterialButton
 import com.google.android.material.card.MaterialCardView
 import mx.yellowme.youst.R
-import mx.yellowme.youst.presentation.showcase.cards.CardAdapter.Companion.MAX_ELEVATION_FACTOR
+import mx.yellowme.youst.showcase.cards.CardAdapter.Companion.MAX_ELEVATION_FACTOR
 import mx.yellowme.youst.core.domain.Challenge
 import mx.yellowme.youst.core.hooks.recycler.ItemListener
 import java.util.*
