+++ app/src/main/java/mx/yellowme/youst/splash/SplashActivity.kt
@@ -1,4 +1,4 @@
-package mx.yellowme.youst.presentation.splash
+package mx.yellowme.youst.splash
 
 import android.os.Bundle
 import kotlinx.android.synthetic.main.activity_splash.*
@@ -6,7 +6,7 @@ import mx.yellowme.youst.R
 import mx.yellowme.youst.core.extensions.launch
 import mx.yellowme.youst.core.hooks.BaseActivity
 import mx.yellowme.youst.core.hooks.animations.setListener
-import mx.yellowme.youst.presentation.showcase.ShowcaseActivity
+import mx.yellowme.youst.showcase.ShowcaseActivity
 
 class SplashActivity : BaseActivity() {
