+++ app/src/main/res/layout/activity_splash.xml
@@ -6,7 +6,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white"
-        tools:context=".presentation.splash.SplashActivity">
+        tools:context=".splash.SplashActivity">
 
     <com.airbnb.lottie.LottieAnimationView
             android:id="@+id/splashAnimationView"
