+++ app/src/main/res/layout/screen_challenge_crazy_lists.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".presentation.challenges.CrazyListsChallengeActivity">
+    tools:context=".challenges.CrazyListsChallengeActivity">
 
     <include layout="@layout/component_back_app_bar"/>
