+++ app/src/main/res/layout/screen_showcase.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".presentation.showcase.ShowcaseActivity">
+    tools:context=".showcase.ShowcaseActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
