+++ app/src/main/res/menu/menu_challenge_help.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".presentation.challenges.CrazyListsChallengeActivity">
+    tools:context=".challenges.CrazyListsChallengeActivity">
 
     <item
         android:id="@+id/action_help"
