+++ core/src/main/java/mx/yellowme/youst/core/data/APIServiceGenerator.kt
@@ -1,4 +1,4 @@
-package mx.yellowme.youst.data;
+package mx.yellowme.youst.core.data;
 
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
