+++ core/src/main/java/mx/yellowme/youst/core/data/PokemonAPIService.kt
@@ -1,4 +1,4 @@
-package mx.yellowme.youst.data;
+package mx.yellowme.youst.core.data;
 
 import mx.yellowme.youst.core.domain.Pokemon
 import retrofit2.Call
